Cambios en formularios, creación registro cliente comun

+++ frontelbuensabor/src/views/AñadirManufacturado.vue
@@ -24,8 +24,7 @@
               />
               <br />
               <b-form-invalid-feedback>
-                <br />Este campo es obligatorio. <br />Recuerde ingresar sólo
-                letras.
+                <br />Este campo es obligatorio. <br />Recuerde ingresar sólo letras.
               </b-form-invalid-feedback>
             </b-form-group>
           </div>
@@ -42,8 +41,7 @@
                 max-rows="6"
               />*
               <b-form-invalid-feedback>
-                <br />Este campo es obligatorio. <br />Recuerde ingresar sólo
-                letras.
+                <br />Este campo es obligatorio. <br />Recuerde ingresar sólo letras.
               </b-form-invalid-feedback>
             </b-form-group>
           </div>
@@ -73,10 +71,8 @@
           <div class="lineaForm">
             <h4 id="datos">*Datos necesarios</h4>
           </div>
-          <div class="lineaFormDerecha" style="float:right">
-            <b-button pill class="boton2" size="md" @click="volver"
-              >Cancelar</b-button
-            >
+          <div class="lineaFormDerecha" style="float: right">
+            <b-button pill class="boton2" size="md" @click="volver">Cancelar</b-button>
             <b-button pill class="boton" size="md" @click.prevent="onSubmit1"
               >Siguiente</b-button
             >
@@ -84,7 +80,7 @@
         </b-form>
       </div>
 
-      <div id="paso2" style="display:none">
+      <div id="paso2" style="display: none">
         <h2>Composición</h2>
         <b-form id="form2">
           <div class="lineaForm">
@@ -106,9 +102,7 @@
                   <b-form-checkbox
                     :key="row.item.idInsumo"
                     v-model="row.item.clicked"
-                    @change="
-                      cambiarEstadoIngrediente($event, row.item.idInsumo)
-                    "
+                    @change="cambiarEstadoIngrediente($event, row.item.idInsumo)"
                   ></b-form-checkbox>
                 </b-form-group>
               </template>
@@ -153,15 +147,13 @@
                 :state="!$v.form2.tiempoCocina.$invalid"
               ></b-form-input>
               <b-form-invalid-feedback>
-                <br />Este campo es obligatorio <br />y sólo admite números
-                mayores a cero <br />sin coma ni punto.
+                <br />Este campo es obligatorio <br />y sólo admite números mayores a cero
+                <br />sin coma ni punto.
               </b-form-invalid-feedback>
             </b-form-group>
           </div>
-          <div class="lineaFormDerecha" style="float:right">
-            <b-button pill class="boton2" size="md" @click="volver"
-              >Cancelar</b-button
-            >
+          <div class="lineaFormDerecha" style="float: right">
+            <b-button pill class="boton2" size="md" @click="volver">Cancelar</b-button>
             <b-button pill class="boton" size="md" @click.prevent="onSubmit2"
               >Siguiente</b-button
             >
@@ -195,9 +187,7 @@
                 ></cantidad>
               </template>
             </b-table>
-            <b-button pill class="boton2" size="md" @click="volver"
-              >Cancelar</b-button
-            >
+            <b-button pill class="boton2" size="md" @click="volver">Cancelar</b-button>
             <b-button pill class="boton" size="md" @click.prevent="onSubmit3"
               >Siguiente</b-button
             >
@@ -216,14 +206,14 @@
               </div>
             </template>
             Todos los datos de esta página son
-            <br />obligatorios para continuar. <br />Recuerde: Las cantidades
-            deben ser sólo números. <br />Verifique los datos.
+            <br />obligatorios para continuar. <br />Recuerde: Las cantidades deben ser
+            sólo números. <br />Verifique los datos.
           </b-toast>
         </b-form>
       </div>
 
       <br />
-      <div id="revision" style="display:none">
+      <div id="revision" style="display: none">
         <h2>Revisión</h2>
         <br />
         <b-form id="formRevision">
@@ -289,17 +279,13 @@
             </b-toast>
             <div class="modalMedida">
               <b-modal ref="modal" hide-footer hide-header centered title>
-                <p class="modalTitulo">
-                  ¡Sugerencia agregada con éxito! Aguarde...
-                </p>
+                <p class="modalTitulo">¡Sugerencia agregada con éxito! Aguarde...</p>
                 <p class="posicion"></p>
               </b-modal>
             </div>
           </div>
           <div class="lineaFormDerecha">
-            <b-button pill class="boton2" size="md" @click="volver"
-              >Cancelar</b-button
-            >
+            <b-button pill class="boton2" size="md" @click="volver">Cancelar</b-button>
             <b-button pill class="boton" size="md" @click.prevent="guardar"
               >Guardar</b-button
             >
@@ -328,8 +314,7 @@ export default {
   mixins: [validationMixin],
   mounted() {
     this.verificarUsuario();
-    this.esNuevo =
-      this.$route.params.id != "undefined" ? !this.esNuevo : this.esNuevo;
+    this.esNuevo = this.$route.params.id != "undefined" ? !this.esNuevo : this.esNuevo;
   },
   props: {
     user: {},
@@ -404,9 +389,7 @@ export default {
         .getOne("manufacturado", this.$route.params.id)
         .then((data) => (this.manufacturado = data))
         .then(() =>
-          this.$route.params.id != "undefined"
-            ? this.completarCamposForm1()
-            : ""
+          this.$route.params.id != "undefined" ? this.completarCamposForm1() : ""
         );
     },
 
@@ -426,13 +409,9 @@ export default {
       this.manufacturado.descripcion = document.getElementById(
         "descripcionManufacturado"
       ).value;
-      this.manufacturado.aptoCeliaco = document.getElementById(
-        "checkbox-1"
-      ).checked;
+      this.manufacturado.aptoCeliaco = document.getElementById("checkbox-1").checked;
       const vegano = document.getElementById("checkbox-2").checked;
-      const vegetariano = vegano
-        ? true
-        : document.getElementById("checkbox-3").checked;
+      const vegetariano = vegano ? true : document.getElementById("checkbox-3").checked;
       this.manufacturado.vegano = vegano;
       this.manufacturado.vegetariano = vegetariano;
       this.siguiente1();
@@ -466,12 +445,10 @@ export default {
     },
 
     cambiarEstadoIngrediente(valor, id) {
-      
       let existe = this.ingredientes.some((i) => {
         return i.idInsumo == id ? true : false;
       });
-      if (existe)
-        valor ? this.agregarIngrediente(id) : this.eliminarIngrediente(id);
+      if (existe) valor ? this.agregarIngrediente(id) : this.eliminarIngrediente(id);
       else if (valor) this.agregarIngrediente(id);
     },
 
@@ -513,9 +490,7 @@ export default {
 
     buscarItem(id, target) {
       if (target == "unidad") {
-        return this.ingredientes.find(
-          (ingrediente) => ingrediente.idInsumo == id
-        );
+        return this.ingredientes.find((ingrediente) => ingrediente.idInsumo == id);
       } else {
         return this.cantidades.find((item) => item.id == id);
       }
@@ -549,7 +524,9 @@ export default {
     onSubmit3() {
       let camposCompletos = this.verificarDatosForm3();
       if (camposCompletos) {
-        this.cantidades.forEach((c) => this.setRecetasNuevas(c.cantidad, c.id));
+        this.cantidades.forEach((c) =>
+          this.setRecetasNuevas(c.cantidad.replace(",", "."), c.id)
+        );
         this.siguiente3();
       } else {
         this.$bvToast.show("toast-datos");
@@ -568,9 +545,7 @@ export default {
 
     verificarDatosForm3() {
       let cantidadesLen = document.getElementsByName("cantidadesInput").length;
-      let rExp = new RegExp(
-        "(^|[ \\t])([-+]?(\\d+|\\.\\d+|\\d+\\.\\d*))($|[^+-.])"
-      );
+      let rExp = new RegExp("(\\d+(?:\\.\\d+)?)");
       let soloNumeros = this.cantidades.every((c) => rExp.test(c.cantidad));
       return (
         soloNumeros &&
@@ -585,8 +560,8 @@ export default {
         return;
       }
       let img = document.getElementById("imagen").files[0];
-      if (img != undefined && img.size / 1024 > 1024) {
-        this.$bvToast.toast(`La imagen no debe superar los 1MB`, {
+      if (img != undefined && img.size / 1024 > 512) {
+        this.$bvToast.toast(`La imagen no debe superar los 512KB`, {
           title: `¡Atención!`,
           toaster: "b-toaster-top-center",
           solid: true,
@@ -630,21 +605,16 @@ export default {
       const formData = new FormData();
       formData.append("file", imagen);
       await axios
-        .post(
-          "http://localhost:9001/buensabor/sugerencia/uploadImg",
-          formData,
-          {
-            headers: {
-              "Content-Type": "multipart/form-data",
-              "Access-Control-Allow-Origins": "*",
-              "cache-control": "no-cache",
-            },
-          }
-        )
+        .post("http://localhost:9001/buensabor/sugerencia/uploadImg", formData, {
+          headers: {
+            "Content-Type": "multipart/form-data",
+            "Access-Control-Allow-Origins": "*",
+            "cache-control": "no-cache",
+          },
+        })
 
         .catch((error) => {
           return error;
-
         });
       return true;
     },
@@ -654,7 +624,6 @@ export default {
         .save("sugerencia", sugerencia)
         .then((data) => (sugerencia = data))
         .catch((error) => {
-
           return error;
         });
       return sugerencia;
